[refactor] Изменил тип для id города в настройках пользователя

+++ app/src/main/java/mr/shtein/buddyandroidclient/model/PersonRequest.kt
@@ -5,7 +5,7 @@ data class PersonRequest(
     val name: String,
     val surname: String,
     val gender: String,
-    val city_id: Int,
+    val city_id: Long,
     val phoneNumber: String,
     val email: String,
     val password: String
